index.html: update note fo contributors/downloads
Rook has precisely one hundred contributors, which does not count as
"hundreds". Rook must reach 200+ contributors for the "hundreds" wording
to have been accurate. This was changed to "a hundred".

There are a recorded 21 million Docker pulls currently, and that now
counts as "tens of millions", up from the previous "millions".

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>