simplify documentation related to finding the current version
We do not need to state the obvious of how to find
the correct version of the documentation. The docs version
is at the top of the page, which is very self-explanatory.
The instructions for where to find the link weren't
even correct since they referred to the bottom left
corner of the page instead of the top right.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>