docs: remove cassandra and nfs doc links from website
Rook.io will only link to the main rook docs for the Ceph
provider. The Cassandra and NFS operators have not been
supported and are fully deprecated, so there is no need
to link to the docs from the website. The docs are still
available if anyone is using them, they are just not
linked from the main site.

Signed-off-by: travisn <tnielsen@redhat.com>